src/requirements.txt
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
dependency-injector==4.38.0
Flask==2.0.1
Flask-WTF==0.15.1
gevent==21.1.2
greenlet==1.1.0
gunicorn==20.0.4
idna==2.10
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
numpy==1.20.3
psycopg2-binary==2.8.6
python-dateutil==2.8.1
python-dotenv==0.19.1
requests==2.25.1
six==1.16.0
urllib3==1.26.5
Werkzeug==2.0.1
WTForms==2.3.3
zope.event==4.5.0
zope.interface==5.4.0
